second commit

+++ src/main/java/com/sample/demo/domains/Child.java
@@ -1,9 +1,10 @@
 package com.sample.demo.domains;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.util.List;
 
 
 @Data
@@ -14,7 +15,7 @@ public class Child {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;
+    private int id;
 
     @Column(name = "parentId")
     private long parentId;
