+++ src/main/java/com/sample/demo/repositories/ParentRepository.java
@@ -3,5 +3,5 @@ package com.sample.demo.repositories;
 import com.sample.demo.domains.Parent;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ParentRepository extends CrudRepository<Parent, Long> {
+public interface ParentRepository extends CrudRepository<Parent, Integer> {
 }
